feat: finish modules
